Updated version with multiple fixes
Fixes:
Issue #2 
Issue #11 
Issue #12 (most or all)

The Google Sheets variant of the toolbox was broken because Google Sheets does not support the INDIRECT function in the data validation setup, while Excel does.
This is fixed by creating a new hidden worksheet that holds the INDIRECT formulas for all the questions and automatically builds named ranges for each question's data validation list. All questions in both the interview and the roadmap (all phases) have been updated to use the named range on the hidden worksheet. Initial testing in Google Sheets has been successful with fixing the data validation and drop-down lists for answers.

Known issues in Google Sheets are in the graphs in the scorecard and roadmap chart.  Google Sheets doesn't honor the custom Axis min/max set from Excel and defaults to an automatic range. This is easily fixed by manually resetting the custom ranges to 0-3.